+++ pkg/interfaces/contracts/liquidity-mining/IGaugeAdder.sol
@@ -13,6 +13,7 @@
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 pragma solidity >=0.7.0 <0.9.0;
+pragma experimental ABIEncoderV2;
 
 import "./IAuthorizerAdaptorEntrypoint.sol";
 import "./IGaugeController.sol";
@@ -21,9 +22,12 @@ import "./ILiquidityGaugeFactory.sol";
 import "./IStakingLiquidityGauge.sol";
 
 interface IGaugeAdder is IAuthentication {
+    // Deprecated. TODO: remove from interfaces, and remove references.
     enum GaugeType { LiquidityMiningCommittee, veBAL, Ethereum, Polygon, Arbitrum, Optimism, Gnosis, ZKSync }
 
-    event GaugeFactoryAdded(GaugeType indexed gaugeType, ILiquidityGaugeFactory gaugeFactory);
+    // String values are hashed when indexed, so we also emit the raw string as a data field for ease of use.
+    event GaugeTypeAdded(string indexed indexedGaugeType, string gaugeType);
+    event GaugeFactorySet(string indexed indexedGaugeType, string gaugeType, ILiquidityGaugeFactory gaugeFactory);
 
     /**
      * @notice Returns the address of the Authorizer adaptor entrypoint contract.
@@ -36,62 +40,52 @@ interface IGaugeAdder is IAuthentication {
     function getGaugeController() external view returns (IGaugeController);
 
     /**
-     * @notice Returns the `index`'th factory for gauge type `gaugeType`
+     * @notice Returns the list of gauge types.
      */
-    function getFactoryForGaugeType(GaugeType gaugeType, uint256 index) external view returns (address);
+    function getGaugeTypes() external view returns (string[] memory);
 
     /**
-     * @notice Returns the number of factories for gauge type `gaugeType`
+     * @notice Returns gauge type name registered at the given index.
      */
-    function getFactoryForGaugeTypeCount(GaugeType gaugeType) external view returns (uint256);
+    function getGaugeTypeAtIndex(uint256 index) external view returns (string memory);
 
     /**
-     * @notice Returns whether `gauge` has been deployed by one of the listed factories for the gauge type `gaugeType`
+     * @notice Returns the number of gauge types.
      */
-    function isGaugeFromValidFactory(address gauge, GaugeType gaugeType) external view returns (bool);
+    function getGaugeTypesCount() external view returns (uint256);
 
     /**
-     * @notice Adds a new gauge to the GaugeController for the "Ethereum" type.
+     * @notice Returns the factory for gauge type `gaugeType`.
      */
-    function addEthereumGauge(IStakingLiquidityGauge gauge) external;
+    function getFactoryForGaugeType(string memory gaugeType) external view returns (ILiquidityGaugeFactory);
 
     /**
-     * @notice Adds a new gauge to the GaugeController for the "Polygon" type.
-     * This function must be called with the address of the *root* gauge which is deployed on Ethereum mainnet.
-     * It should not be called with the address of the gauge which is deployed on Polygon
+     * @notice Returns true if `gauge` has been deployed by the factory for the gauge type `gaugeType`.
+     * Note that if a gauge type's factory changes then this function will start returning false for previously
+     * valid gauges.
      */
-    function addPolygonGauge(address rootGauge) external;
+    function isGaugeFromValidFactory(address gauge, string memory gaugeType) external view returns (bool);
 
     /**
-     * @notice Adds a new gauge to the GaugeController for the "Arbitrum" type.
-     * This function must be called with the address of the *root* gauge which is deployed on Ethereum mainnet.
-     * It should not be called with the address of the gauge which is deployed on Arbitrum
+     * @notice Adds a new `gaugeType` corresponding to a new network, which allows setting a factory and adding gauges
+     * for the type later on.
+     * @param gaugeType Name of the new gauge type.
      */
-    function addArbitrumGauge(address rootGauge) external;
+    function addGaugeType(string memory gaugeType) external;
 
     /**
-     * @notice Adds a new gauge to the GaugeController for the "Optimism" type.
-     * This function must be called with the address of the *root* gauge which is deployed on Ethereum mainnet.
-     * It should not be called with the address of the gauge which is deployed on Optimism.
+     * @notice Adds a new gauge to the GaugeController for the given `gaugeType` type.
+     * @dev When adding gauges for L2 networks or sidechains, this must be called with the address of the *root* gauge
+     * which is deployed on Ethereum. It should *not* be called with the address of the child gauge which is deployed on
+     * the L2 / sidechain.
+     *
+     * If the gauge added is an Ethereum gauge, it cannot be a gauge for the 80BAL-20WETH pool.
      */
-    function addOptimismGauge(address rootGauge) external;
+    function addGauge(address gauge, string memory gaugeType) external;
 
     /**
-     * @notice Adds a new gauge to the GaugeController for the "Gnosis" type.
-     * This function must be called with the address of the *root* gauge which is deployed on Ethereum mainnet.
-     * It should not be called with the address of the gauge which is deployed on Gnosis Chain.
+     * @notice Sets `factory` as the allowlisted factory contract for gauges with type `gaugeType`.
+     * @dev This function can either set a new factory or replace an existing one.
      */
-    function addGnosisGauge(address rootGauge) external;
-
-    /**
-     * @notice Adds a new gauge to the GaugeController for the "ZKSync" type.
-     * This function must be called with the address of the *root* gauge which is deployed on Ethereum mainnet.
-     * It should not be called with the address of the gauge which is deployed on ZKSync.
-     */
-    function addZKSyncGauge(address rootGauge) external;
-
-    /**
-     * @notice Adds `factory` as an allowlisted factory contract for gauges with type `gaugeType`.
-     */
-    function addGaugeFactory(ILiquidityGaugeFactory factory, GaugeType gaugeType) external;
+    function setGaugeFactory(ILiquidityGaugeFactory factory, string memory gaugeType) external;
 }