+++ pkg/interfaces/CHANGELOG.md
@@ -9,6 +9,14 @@
 - Added `IVotingEscrowRemapper`.
 - Added `IOmniVotingEscrowAdaptor` and `IOmniVotingEscrowAdaptorSettings`.
 
+### Breaking changes
+
+- Refactored `IGaugeAdder`
+  - Gauge types changed from enum to string across the interface.
+  - Removed specific `add<Network>Gauge` in favor of a single `addGauge` method with a type argument.
+  - Added event `GaugeTypeAdded`, and changed `GaugeFactoryAdded` to `GaugeFactorySet`.
+  - Added `getGaugeTypes` method.
+
 ## 0.4.0 (2023-03-15)
 
 - Added `IERC4626`.